requirements.txt
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
faiss-cpu
fastapi
uvicorn
python-multipart
pydantic
streamlit
requests
faster-whisper
sounddevice
scipy
git+https://github.com/openai/CLIP.git